hermes_translator/requirements.txt
argostranslate==1.9.6
async-timeout==4.0.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
ctranslate2==4.3.1
filelock==3.15.1
fsspec==2024.6.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
lingua-language-detector==2.0.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
packaging==24.1
protobuf==5.27.1
PyYAML==6.0.1
redis==5.0.6
regex==2024.5.15
requests==2.32.3
sacremoses==0.0.53
sentencepiece==0.2.0
six==1.16.0
stanza==1.1.1
sympy==1.12.1
torch==2.3.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.1
